Add files via upload
Delving into the realm of COVID-19 data, this project embarks on a journey of exploration, cleaning, and forecasting, utilizing the power of SQL Server and Tableau. Together, we embark on a quest to uncover patterns and predict the pandemic's trajectory for the next two years. Our journey unveils enlightening insights, revealing the distribution of deaths across continents, the impact of infection on countries' populations, and the global death percentage. Embark with us as we illuminate the COVID-19 landscape through data-driven storytelling